* commit '4e7bb65454122ee1ac66a8d7a0a1f88f1b39c6bc':
  SHIELD-137: remove trace
  SHIELD-137: clean-up
  SHIELD-137: improve package name & type parsing, plus new checks

+++ generate_shield_package.sh
@@ -7,24 +7,35 @@ error() {
   exit 0
 }
 
+[[ ! -f $p_path ]] && error "Could not find package in file system"
+
 p_name=$(basename $p_path)
-p_id="${p_name%.tar.gz}"
+p_id=$p_name
 
 [[ -z $p_id  || -z $p_path ]] && error "The path to an OSM package must be provided"
 
-p_type="${p_id##*_}"
-p_type="${p_type%.*}"
-p_id_s="${p_id%_*}"
+# Apply regexp and capture name and type for package
+if [[ "$p_id" =~ ([a-zA-Z0-9]{1,})_([a-zA-Z0-9]{1,})[_(?:.*)]* ]]; then
+  p_id_s=${BASH_REMATCH[1]}
+  p_type=${BASH_REMATCH[2]}
+  # Replace OSM-specific naming
+  p_type="${p_type/vnfd/vnf}"
+  p_type="${p_type/nsd/ns}"
+fi
 
 [[ -z $p_id_s ]] && error "Could not determine name of the package"
 [[ -z $p_type ]] && error "Could not determine package type (NS or VNF). Expected formats: \${pkg_name}_ns.tar.gz, \${pkg_name}_vnf.tar.gz"
 
 repo_root=$PWD
+
+manifest_path="${repo_root}/security-manifest/${p_type}/${p_id_s}/manifest.yaml"
+[[ ! -f $manifest_path ]] && error "Could not find SHIELD security manifest for this package"
+
 pkg_tmp=$(mktemp -d)
 
 # Copy package and corresponding security manifest
 cp -Rp $p_path $pkg_tmp/
-cp -p security-manifest/${p_type}/${p_id_s}/manifest.yaml $pkg_tmp/
+cp -p $manifest_path $pkg_tmp/
 cd $pkg_tmp
 
 juju_in=$(dpkg -l | grep juju)
@@ -34,15 +45,10 @@ fi
 
 clear
 
-# Uncompress package to verify contents
-#tar -zxf $p_name
-#rm $p_id
-
 gen_hash_for_pkg() {
   ns_pkg=$(sed -n -e '/package/ s/.*\: *//p' manifest.yaml)
   sha256_p_ns=$(sha256sum $ns_pkg)
   sha256_p_ns=$(echo $sha256_p_ns | awk '{print $1;}')
-#  grep -l "hash: <sha256-based hash of the .tar.gz package defined above>" manifest.yaml | xargs sed -i -e "s|hash: <sha256-based hash of the .tar.gz package defined above>|hash: $sha256_p_ns|"
   sed -i -e "s/hash: <sha256-based hash of the .tar.gz package defined above>/hash: $sha256_p_ns/g" manifest.yaml
 }
